add docstrings, move and update test
