+++ python/cugraph-service/client/cugraph_service_client.egg-info/SOURCES.txt
@@ -1,0 +1,17 @@
+setup.cfg
+setup.py
+cugraph_service_client/__init__.py
+cugraph_service_client/_version.py
+cugraph_service_client/client.py
+cugraph_service_client/cugraph_service_thrift.py
+cugraph_service_client/defaults.py
+cugraph_service_client/exceptions.py
+cugraph_service_client/remote_graph.py
+cugraph_service_client/remote_graph_utils.py
+cugraph_service_client/types.py
+cugraph_service_client.egg-info/PKG-INFO
+cugraph_service_client.egg-info/SOURCES.txt
+cugraph_service_client.egg-info/dependency_links.txt
+cugraph_service_client.egg-info/requires.txt
+cugraph_service_client.egg-info/top_level.txt
+cugraph_service_client.egg-info/zip-safe
