+++ python/cugraph-service/client/cugraph_service_client.egg-info/requires.txt
@@ -1,0 +1,1 @@
+thriftpy2
