+++ python/cugraph-service/client/cugraph_service_client.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+cugraph_service_client
