+++ python/cugraph-service/client/cugraph_service_client/client.py
@@ -33,7 +33,22 @@ from cugraph_service_client.cugraph_service_thrift import create_client
 
 
 class RunAsyncioThread(threading.Thread):
+    """
+    This class provides a thread whose purpose is to start a new
+    event loop and call the provided function.
+    """
+
     def __init__(self, func, args, kwargs):
+        """
+        Parameters
+        ----------
+        func : function
+            The function that will be run.
+        *args : args
+            The arguments to the given function.
+        **kwargs : kwargs
+            The keyword arguments to the given function.
+        """
         self.func = func
         self.args = args
         self.kwargs = kwargs
@@ -41,10 +56,38 @@ class RunAsyncioThread(threading.Thread):
         super().__init__()
 
     def run(self):
+        """
+        Runs this thread's previously-provided function inside
+        a new event loop.  Returns the result.
+
+        Returns
+        -------
+        The returned object of the previously-provided function.
+        """
         self.result = asyncio.run(self.func(*self.args, **self.kwargs))
 
 
 def run_async(func, *args, **kwargs):
+    """
+    If no loop is running on the current thread,
+    this method calls func using a new event
+    loop using asyncio.run.  If a loop is running, this
+    method starts a new thread, and calls func on a new
+    event loop in the new thread.
+
+    Parameters
+    ----------
+    func : function
+        The function that will be run.
+    *args : args
+        The arguments to the given function.
+    **kwargs : kwargs
+        The keyword arguments to the given function.
+
+    Returns
+    -------
+    The output of the given function.
+    """
     try:
         loop = asyncio.get_running_loop()
     except RuntimeError:
