+++ python/cugraph-service/server/cugraph_service_server.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+setup.cfg
+setup.py
+cugraph_service_server/__init__.py
+cugraph_service_server/__main__.py
+cugraph_service_server/_version.py
+cugraph_service_server/cugraph_handler.py
+cugraph_service_server.egg-info/PKG-INFO
+cugraph_service_server.egg-info/SOURCES.txt
+cugraph_service_server.egg-info/dependency_links.txt
+cugraph_service_server.egg-info/entry_points.txt
+cugraph_service_server.egg-info/requires.txt
+cugraph_service_server.egg-info/top_level.txt
+cugraph_service_server.egg-info/zip-safe
