+++ python/cugraph-service/server/cugraph_service_server.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+cugraph-service-server = cugraph_service_server.__main__:main
