+++ python/cugraph-service/server/cugraph_service_server.egg-info/requires.txt
@@ -1,0 +1,8 @@
+cugraph-service-client
+cugraph
+cupy<12.0.0a0,>=9.5.0
+numpy
+ucx-py
+distributed>=2022.9.2
+dask-cuda
+thriftpy2
