+++ python/cugraph-service/server/cugraph_service_server.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+cugraph_service_server
