+++ python/cugraph-service/tests/test_e2e.py
@@ -515,29 +515,6 @@ def test_uniform_neighbor_sampling(client_with_edgelist_csv_loaded):
     )
 
 
-def test_inside_asyncio_event_loop(client_with_edgelist_csv_loaded):
-    import asyncio
-    from cugraph_service_client import defaults
-
-    client, test_data = client_with_edgelist_csv_loaded
-
-    start_list = [1, 2, 3]
-    fanout_vals = [2, 2, 2]
-    with_replacement = True
-
-    async def uns():
-        return client.uniform_neighbor_sample(
-            start_list=start_list,
-            fanout_vals=fanout_vals,
-            with_replacement=with_replacement,
-            graph_id=defaults.graph_id,
-            result_device=0,
-        )
-
-    # ensure call succeeds; have confirmed this fails without fix in client
-    assert asyncio.run(uns()) is not None
-
-
 def test_create_property_graph(client):
     old_ids = set(client.get_graph_ids())
     pG = client.graph()
